- few bugs resolved. - di3 disk occupancy and cache are optimized.

+++ Di3/Interfaces/IMetaData.cs
@@ -12,6 +12,6 @@ namespace Interfaces
         C left { set; get; }
         C right { set; get; }
         double value { set; get; }
-        UInt64 hashKey { set; get; }
+        UInt32 hashKey { set; get; }
     }
 }
